feat paciente, fix id long

+++ src/main/java/com/atendimentomedico/crudmedico/service/PacienteService.java
@@ -21,7 +21,7 @@ public class PacienteService {
         return repository.findAll();
     }
 
-    public Paciente getPacienteById(Integer id) {
+    public Paciente getPacienteById(Long id) {
         return repository.findById(id).orElse(null);
     }
